+++ CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this project will be documented in this file.
 - Replace `lazy_static` with `std::cell::LazyCell` ([#604]).
 - Promote Druid `30.0.0` to LTS, deprecate `26.0.0` ([#631]).
 
+### Fixed
+
+- BREAKING: The fields `connection` and `host` on `S3Connection` as well as `bucketName` on `S3Bucket`are now mandatory ([#632]).
+
 ### Removed
 
 - test: Remove ZooKeeper 3.8.4 ([#621]).
@@ -22,6 +26,7 @@ All notable changes to this project will be documented in this file.
 [#604]: https://github.com/stackabletech/druid-operator/pull/604
 [#621]: https://github.com/stackabletech/druid-operator/pull/621
 [#631]: https://github.com/stackabletech/druid-operator/pull/631
+[#632]: https://github.com/stackabletech/druid-operator/pull/632
 
 ## [24.7.0] - 2024-07-24